backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vincent Tiu <vincetiu8@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask-sqlalchemy = "^3.0.3"
python-dotenv = "^1.0.0"
pyjwt = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

